requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
certifi==2020.11.8        # via requests
cffi==1.14.3              # via cryptography
chardet==3.0.4            # via requests
click-log==0.3.2          # via -r requirements.in
click==7.1.2              # via -r requirements.in, click-log, flask
cryptography==3.2.1       # via msal, pyjwt
flask==1.1.2              # via -r requirements.in
idna==2.10                # via requests
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
markupsafe==1.1.1         # via jinja2
msal==1.6.0               # via -r requirements.in
pycparser==2.20           # via cffi
pyjwt[crypto]==1.7.1      # via msal
python-dateutil==2.8.1    # via -r requirements.in
python-dotenv==0.15.0     # via -r requirements.in
requests==2.25.0          # via -r requirements.in, msal
six==1.15.0               # via cryptography, python-dateutil
urllib3==1.26.2           # via requests
werkzeug==1.0.1           # via flask
